Add interop tests between different feature sets
