+++ mls-rs/src/group/proposal_filter/filtering_lite.rs
@@ -122,11 +122,17 @@ where
             )
             .await?;
 
+        let new_context_extensions = proposals
+            .group_context_extensions
+            .first()
+            .map(|gce| gce.proposal.clone());
+
         Ok(ApplyProposalsOutput {
             new_tree,
             indexes_of_added_kpkgs: added,
             #[cfg(feature = "external_commit")]
             external_init_index: None,
+            new_context_extensions,
         })
     }
