+++ mls-rs/test_harness_integration/Cargo.toml
@@ -6,13 +6,24 @@ edition = "2021"
 publish = false
 
 [dependencies]
-mls-rs = { path = "..", features = ["external_client", "rfc_compliant"]}
+mls-rs = { path = "..", default-features = false, features = ["std", "external_client", "external_commit", "state_update", "all_extensions"]}
 tonic = "0.7"
 prost = "0.10"
 tokio = { version = "1", features = ["macros", "rt-multi-thread"] }
 clap = { version = "3", features = ["derive"] }
 thiserror = "1"
 futures = "0.3.25"
+hex = "0.4"
+
+[features]
+tree_index = [ "mls-rs/tree_index" ]
+private_message = [ "mls-rs/private_message" ]
+prior_epoch = [ "mls-rs/prior_epoch" ]
+out_of_order = [ "mls-rs/out_of_order", "private_message" ]
+psk = [ "mls-rs/psk" ]
+custom_proposal = [ "mls-rs/custom_proposal" ]
+by_ref_proposal = [ "mls-rs/by_ref_proposal", "mls-rs/external_proposal" ]
+default = ["tree_index", "private_message", "prior_epoch", "out_of_order", "psk", "custom_proposal", "by_ref_proposal"]
 
 [target.'cfg(target_arch = "wasm32")'.dependencies]
 mls-rs-crypto-rustcrypto = { path = "../../mls-rs-crypto-rustcrypto", features = ["browser"], version = "*" }
