+++ mls-rs/test_harness_integration/configs/application_out_of_order_within_epoch.json
@@ -1,0 +1,16 @@
+{
+  "scripts": {
+    "out_of_order_within_epoch": [
+      {"action": "createGroup", "actor": "alice"},
+      {"action": "externalJoin", "actor": "alice", "joiner": "bob"},
+      
+      {"action": "protect", "actor": "alice", "authenticatedData": "btw1", "plaintext": "hello world 1"},
+      {"action": "protect", "actor": "alice", "authenticatedData": "btw2", "plaintext": "hello world 2"},
+      {"action": "protect", "actor": "alice", "authenticatedData": "btw3", "plaintext": "hello world 3"},
+
+      {"action": "unprotect", "actor": "bob", "ciphertext": 4 },
+      {"action": "unprotect", "actor": "bob", "ciphertext": 3 },
+      {"action": "unprotect", "actor": "bob", "ciphertext": 2 }
+    ]
+  }
+}
