+++ mls-rs/test_harness_integration/configs/commit_with_psk.json
@@ -1,0 +1,133 @@
+{
+  "scripts": {
+    "external_psk": [
+      {"action": "createGroup", "actor": "alice"},
+      {"action": "externalJoin", "actor": "alice", "joiner": "bob"},
+      {"action": "installExternalPSK", "clients": ["alice", "bob"]},
+      {"action": "installExternalPSK", "clients": ["alice", "bob"]},
+      {"action": "installExternalPSK", "clients": ["alice", "bob"]},
+
+      {"action": "externalPSKProposal", "pskID": 2, "actor": "alice"},
+      {"action": "externalPSKProposal", "pskID": 3, "actor": "bob"},
+      {
+        "action": "fullCommit",
+        "actor": "alice",
+        "byReference": [5, 6],
+        "byValue": [{"proposalType": "externalPSK", "pskID": 4}],
+        "members": ["bob"],
+        "force_path": true
+      },
+
+      {"action": "externalPSKProposal", "pskID": 2, "actor": "alice"},
+      {"action": "externalPSKProposal", "pskID": 3, "actor": "bob"},
+      {
+        "action": "fullCommit",
+        "actor": "alice",
+        "byReference": [8, 9],
+        "byValue": [{"proposalType": "externalPSK", "pskID": 4}],
+        "members": ["bob"],
+        "force_path": false
+      }
+    ],
+
+    "resumption_psk": [
+      {"action": "createGroup", "actor": "alice"},
+      {"action": "externalJoin", "actor": "alice", "joiner": "bob"},
+      {"action": "fullCommit", "actor": "alice", "members": ["bob"]},
+      {"action": "fullCommit", "actor": "alice", "members": ["bob"]},
+
+      {"action": "resumptionPSKProposal", "epochID": 1, "actor": "alice"},
+      {"action": "resumptionPSKProposal", "epochID": 2, "actor": "bob"},
+      {
+        "action": "fullCommit",
+        "actor": "alice",
+        "byReference": [4, 5],
+        "byValue": [{"proposalType": "resumptionPSK", "epochID": 3}],
+        "members": ["bob"],
+        "force_path": true
+      },
+
+      {"action": "resumptionPSKProposal", "epochID": 2, "actor": "alice"},
+      {"action": "resumptionPSKProposal", "epochID": 3, "actor": "bob"},
+      {
+        "action": "fullCommit",
+        "actor": "alice",
+        "byReference": [7, 8],
+        "byValue": [{"proposalType": "resumptionPSK", "epochID": 4}],
+        "members": ["bob"],
+        "force_path": true
+      }
+    ],
+
+    "all_together_alice_proposes": [
+      {"action": "createGroup", "actor": "alice"},
+      {"action": "externalJoin", "actor": "alice", "joiner": "bob"},
+      {"action": "createKeyPackage", "actor": "charlie1"},
+      {"action": "createKeyPackage", "actor": "charlie2"},
+      {"action": "addProposal", "actor": "alice", "keyPackage": 2},
+      {"action": "addProposal", "actor": "alice", "keyPackage": 3},
+      {"action": "fullCommit", "actor": "alice", "byReference": [4, 5], "members": ["bob"], "joiners": ["charlie1", "charlie2"]},
+      {"action": "installExternalPSK", "clients": ["alice", "bob"]},
+      {"action": "installExternalPSK", "clients": ["alice", "bob"]},
+
+      {"action": "createKeyPackage", "actor": "charlie3"},
+      {"action": "addProposal", "actor": "alice", "keyPackage": 9},
+      {"action": "removeProposal", "actor": "alice", "removed": "charlie1"},
+      {"action": "groupContextExtensionsProposal", "actor": "alice", "extensions": [
+        {"extension_type": 3, "extension_data": "AAAA"},
+        {"extension_type": 5, "extension_data": "AA=="}
+      ]},
+      {"action": "externalPSKProposal", "pskID": 7, "actor": "alice"},
+      {"action": "resumptionPSKProposal", "epochID": 1, "actor": "alice"},
+      {"action": "createKeyPackage", "actor": "charlie4"},
+      {
+        "action": "fullCommit",
+        "actor": "alice",
+        "byReference": [10, 11, 12, 13, 14],
+        "byValue": [
+          {"proposalType": "add", "keyPackage": 15},
+          {"proposalType": "remove", "removed": "charlie2"},
+          {"proposalType": "externalPSK", "pskID": 8},
+          {"proposalType": "resumptionPSK", "epochID": 2}
+        ],
+        "members": ["bob"]
+      }
+    ],
+
+    "all_together_bob_proposes": [
+      {"action": "createGroup", "actor": "alice"},
+      {"action": "externalJoin", "actor": "alice", "joiner": "bob"},
+      {"action": "createKeyPackage", "actor": "charlie1"},
+      {"action": "createKeyPackage", "actor": "charlie2"},
+      {"action": "addProposal", "actor": "alice", "keyPackage": 2},
+      {"action": "addProposal", "actor": "alice", "keyPackage": 3},
+      {"action": "fullCommit", "actor": "alice", "byReference": [4, 5], "members": ["bob"], "joiners": ["charlie1", "charlie2"]},
+      {"action": "installExternalPSK", "clients": ["alice", "bob"]},
+      {"action": "installExternalPSK", "clients": ["alice", "bob"]},
+
+      {"action": "createKeyPackage", "actor": "charlie3"},
+      {"action": "addProposal", "actor": "bob", "keyPackage": 9},
+      {"action": "removeProposal", "actor": "bob", "removed": "charlie1"},
+      {"action": "groupContextExtensionsProposal", "actor": "bob", "extensions": [
+        {"extension_type": 3, "extension_data": "AAAA"},
+        {"extension_type": 5, "extension_data": "AA=="}
+      ]},
+      {"action": "externalPSKProposal", "pskID": 7, "actor": "bob"},
+      {"action": "resumptionPSKProposal", "epochID": 1, "actor": "bob"},
+      {"action": "updateProposal", "actor": "bob"},
+      {"action": "createKeyPackage", "actor": "charlie4"},
+      {
+        "action": "fullCommit",
+        "actor": "alice",
+        "byReference": [10, 11, 12, 13, 14, 15],
+        "byValue": [
+          {"proposalType": "add", "keyPackage": 16},
+          {"proposalType": "remove", "removed": "charlie2"},
+          {"proposalType": "externalPSK", "pskID": 8},
+          {"proposalType": "resumptionPSK", "epochID": 2}
+        ],
+        "members": ["bob"]
+      }
+    ]
+  }
+}
