+++ mls-rs/test_harness_integration/configs/external_proposals.json
@@ -1,0 +1,69 @@
+{
+  "scripts": {
+    "joiner_signed_add": [
+      {"action": "createGroup", "actor": "alice"},
+      {"action": "newMemberAddProposal", "actor": "alice", "joiner": "bob" },
+      {"action": "fullCommit", "actor": "alice", "byReference": [1], "joiners": ["bob"] }
+    ],
+
+    "external_add": [
+      {"action": "createGroup", "actor": "alice"},
+      {"action": "externalJoin", "actor": "alice", "joiner": "bob"},
+
+      {"action": "addExternalSigner", "actor": "alice", "signer": "ds"},
+      {"action": "fullCommit", "actor": "alice", "byReference": [2], "members": ["bob"]},
+
+      {"action": "createKeyPackage", "actor": "charlie" },
+      {"action": "externalSignerProposal", "actor": "ds", "member": "alice",
+       "description": { "proposalType": "add", "keyPackage": 4 } },
+      {"action": "fullCommit", "actor": "alice", "byReference": [5], "members": ["bob"], 
+       "joiners": ["charlie"] }
+    ],
+
+    "external_remove": [
+      {"action": "createGroup", "actor": "alice"},
+      {"action": "externalJoin", "actor": "alice", "joiner": "bob"},
+      {"action": "externalJoin", "actor": "alice", "joiner": "charlie", "members": ["bob"]},
+
+      {"action": "addExternalSigner", "actor": "alice", "signer": "ds"},
+      {"action": "fullCommit", "actor": "alice", "byReference": [3], "members": ["bob", "charlie"]},
+
+      {"action": "externalSignerProposal", "actor": "ds", "member": "alice",
+       "description": { "proposalType": "remove", "removed": "bob" } },
+      {"action": "fullCommit", "actor": "alice", "byReference": [5], "members": ["charlie"] }
+    ],
+
+    "group_context_extensions": [
+      {"action": "createGroup", "actor": "alice"},
+      {"action": "externalJoin", "actor": "alice", "joiner": "bob"},
+      {"action": "externalJoin", "actor": "alice", "joiner": "charlie", "members": ["bob"]},
+
+      {"action": "addExternalSigner", "actor": "alice", "signer": "ds"},
+      {"action": "fullCommit", "actor": "alice", "byReference": [3], "members": ["bob", "charlie"]},
+      
+      {"action": "externalSignerProposal", "actor": "ds", "member": "alice", "description": {
+        "proposalType": "groupContextExtensions", "extensions": [ {"extension_type": 3, "extension_data": "AAAA"} ] } },
+      {"action": "fullCommit", "actor": "alice", "byReference": [5], "members": ["bob", "charlie"] }
+    ],
+    
+    "multiple_external": [
+      {"action": "createGroup", "actor": "alice"},
+      {"action": "externalJoin", "actor": "alice", "joiner": "bob"},
+      {"action": "externalJoin", "actor": "alice", "joiner": "charlie", "members": ["bob"]},
+      {"action": "createKeyPackage", "actor": "dave" },
+      {"action": "createKeyPackage", "actor": "edward" },
+
+      {"action": "addExternalSigner", "actor": "alice", "signer": "ds1"},
+      {"action": "fullCommit", "actor": "alice", "byReference": [5], "members": ["bob", "charlie"]},
+      
+      {"action": "addExternalSigner", "actor": "alice", "signer": "ds2"},
+      {"action": "fullCommit", "actor": "alice", "byReference": [7], "members": ["bob", "charlie"]},
+      
+      {"action": "externalSignerProposal", "actor": "ds1", "member": "alice",
+       "description": { "proposalType": "add", "keyPackage": 3 } },
+      {"action": "externalSignerProposal", "actor": "ds1", "member": "alice",
+       "description": { "proposalType": "add", "keyPackage": 4 } },
+      {"action": "fullCommit", "actor": "alice", "byReference": [9, 10], "members": ["bob", "charlie"] }
+    ]
+  }
+}
